patch: Automatically determine patch strip level
This allows us to directly use patches generated using
`git format-patch` and also those authored by external
entities without needing to add metadata (although this
may not be a bad idea).

Salvaging something from the time spent on `patch.py`
(which I intend to see through at some time) though it
is a genuinely useful feature that will save time and
make creating and upstreaming patches a little easier.

Also added a test for this new feature and a test for
`patch` itself.